requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
boto3
ffmpeg-python
